Enable local to local synchronisation
When neither the first nor the last arguments are a remote file
name, we are copying from and to the local machine.  In this case, the
current process will send the file corresponding to the first argument
to a new process that will receive the file corresponding to the second
argument.