Handle destination directories
When the destination file exist and is a directory, synchronize a file
with the same name as the source file in this destination directory.

```
radjust user@host:file /tmp
```

Will create `/tmp/file`.

Only the basename of the transfered file is now sent (it used to be the
full name on the remote host), reducing the amount of data transceived
in the test suite by 5 bytes ("/tmp/" being stripped).